+++ src/types/components/popover/theme.d.ts
@@ -1,0 +1,8 @@
+declare const popover: {
+    base: string;
+    anchor: string;
+    content: string;
+    wrapper: string;
+};
+export type PopoverUi = typeof popover;
+export default popover;